[KGEN] Fix sign comparison warning in `ZAPOps.cpp` (#3583)
There is a sign comparison warning:

```
modular/KGEN/lib/ZAPDialect/ZAPOps.cpp:209:22: warning: comparison of integers of different signs: 'int64_t' (aka 'long long') and 'size_t' (aka 'unsigned long') [-Wsign-compare]
    if (dim.getInt() >= tensorType.getRank())
```

Fix the sign comparison warning by casting the result of `dim.getInt()`
to a `size_t` instead so we compare two types of type `unsigned long`.

MODULAR_ORIG_COMMIT_REV_ID: f445e2c097b4fc36a74e83589f6e2f98c309eedd

+++ KGEN/lib/ZAPDialect/ZAPOps.cpp
@@ -206,7 +206,7 @@ OpFoldResult TensorDimOp::fold(ArrayRef<Attribute> operands) {
   if (auto dim = dyn_cast_if_present<IntegerAttr>(operands[1])) {
     auto tensorType = getTensor().getType();
     // If the index is out of bounds, then we do not fold the op.
-    if (dim.getInt() >= tensorType.getRank())
+    if (static_cast<size_t>(dim.getInt()) >= tensorType.getRank())
       return {};
     return tensorType.getShape().data()[dim.getInt()];
   }
